Refactor bitcoin startup
