+++ src/lib.rs
@@ -1,2 +1,3 @@
+pub mod bitcoin;
 pub mod create_comit_app;
 pub mod ethereum;
